+++ initdb.d/02-create-db.sql
@@ -1,0 +1,1 @@
+create database wil_dev owner postgres;