+++ resources/migrations/20220930050830-weeks.up.sql
@@ -1,0 +1,5 @@
+CREATE TABLE weeks (
+  id    SERIAL PRIMARY KEY,
+  klass VARCHAR(10),
+  date  DATE
+);