+++ src/clj/wil/routes/home.clj
@@ -1,21 +1,56 @@
 (ns wil.routes.home
   (:require
+   [buddy.hashers :as hashers]
+   [clojure.tools.logging :as log]
+   [hato.client :as hc]
+   [ring.util.http-response :as response]
    [wil.layout :as layout]
-   [wil.db.core :as db]
-   [clojure.java.io :as io]
-   [wil.middleware :as middleware]
-   [ring.util.response]
-   [ring.util.http-response :as response]))
+   [wil.middleware :as middleware]))
 
-(defn home-page [request]
-  (layout/render request "home.html"))
+(defn home-page
+  [request]
+  (if (get-in request [:session :identity])
+    (layout/render request "home.html")
+    (response/found "/login")))
+
+(def api-user "https://l22.melt.kyutech.ac.jp/api/user/")
+(defn get-user
+  "retrieve str login's info from API.
+   note: parameter is a string. cf. (db/get-user {:login login})"
+  [login]
+  (let [ep   (str api-user login)
+        resp (hc/get ep {:as :json})]
+    (log/info "login" (get-in resp [:body :login]))
+    (:body resp)))
+
+(defn login-page [request]
+  (layout/render request "login.html" {:flash (:flash request)}))
+
+(defn login-post [{{:keys [login password klass]} :params}]
+  (let [user (get-user login)]
+    (if (and (seq user)
+             (= (:login user) login)
+             (hashers/check password (:password user)))
+      (do
+        (log/info "login success" login)
+        (-> (response/found "/")
+            (assoc-in [:session :identity] login)
+            (assoc-in [:session :klass] klass)))
+      (do
+        (log/info "login faild" login)
+        (-> (response/found "/login")
+            (assoc :flash "login failure"))))))
+
+(defn logout [_]
+  (-> (response/found "/login")
+      (assoc :session {})))
 
 (defn home-routes []
   [""
    {:middleware [middleware/wrap-csrf
                  middleware/wrap-formats]}
    ["/" {:get home-page}]
-   ["/docs" {:get (fn [_]
-                    (-> (response/ok (-> "docs/docs.md" io/resource slurp))
-                        (response/header "Content-Type" "text/plain; charset=utf-8")))}]])
+   ["/login" {:get  login-page
+              :post login-post}]
+   ["/logout" {:get logout}]])