+++ src/cljs/wil/core.cljs
@@ -1,15 +1,17 @@
 (ns wil.core
   (:require
-    [reagent.core :as r]
-    [reagent.dom :as rdom]
-    [goog.events :as events]
-    [goog.history.EventType :as HistoryEventType]
-    [markdown.core :refer [md->html]]
-    [wil.ajax :as ajax]
-    [ajax.core :refer [GET POST]]
-    [reitit.core :as reitit]
-    [clojure.string :as string])
+   [reagent.core :as r]
+   [reagent.dom :as rdom]
+   [goog.events :as events]
+   [goog.history.EventType :as HistoryEventType]
+   [markdown.core :refer [md->html]]
+   [wil.ajax :as ajax]
+   [ajax.core :refer [GET POST]]
+   [reitit.core :as reitit]
+   [clojure.string :as string])
   (:import goog.History))
+
+(def ^:private version "0.2.0")
 
 (defonce session (r/atom {:page :home}))
 
@@ -19,31 +21,63 @@
     :class (when (= page (:page @session)) "is-active")}
    title])
 
-(defn navbar [] 
+(defn navbar []
   (r/with-let [expanded? (r/atom false)]
     [:nav.navbar.is-info>div.container
      [:div.navbar-brand
-      [:a.navbar-item {:href "/" :style {:font-weight :bold}} "wil"]
+      [:a.navbar-item {:href "/" :style {:font-weight :bold}} "WIL"]
       [:span.navbar-burger.burger
        {:data-target :nav-menu
         :on-click #(swap! expanded? not)
         :class (when @expanded? :is-active)}
-       [:span][:span][:span]]]
+       [:span] [:span] [:span]]]
      [:div#nav-menu.navbar-menu
       {:class (when @expanded? :is-active)}
       [:div.navbar-start
-       [nav-link "#/" "Home" :home]
-       [nav-link "#/about" "About" :about]]]]))
+       #_[nav-link "#/" "Home" :home]
+       [nav-link "https://py99.melt.kyutech.ac.jp" "Py99"]
+       [nav-link "https://qa.melt.kyutech.ac.jp" "QA"]
+       [nav-link "#/about" "About" :about]
+       [nav-link "/logout" "Logout"]]]]))
 
 (defn about-page []
   [:section.section>div.container>div.content
-   [:img {:src "/img/warning_clojure.png"}]])
+   [:img {:src "/img/warning_clojure.png"}]
+   [:p "version " version]])
 
+(defn notes-component []
+  [:div])
+
+(defn today-is-klass-day?
+  [klass]
+  true)
+
+(defn send-note
+  [login text]
+  (js/alert (str "send-note " login " "(subs text 0 10))))
+
+(defonce note (r/atom ""))
+
+(defn new-note-compoment []
+  [:div
+   [:div
+    [:textarea
+     {:id "note"
+      :value @note
+      :on-change #(reset! note (-> % .-target .-value))}]]
+   [:div
+    [:button
+     {:on-click #(send-note js/login @note)
+      :class "button is-primary"}
+     "submit"]]])
 
 (defn home-page []
   [:section.section>div.container>div.content
-   (when-let [docs (:docs @session)]
-     [:div {:dangerouslySetInnerHTML {:__html (md->html docs)}}])])
+   [:h3 js/login " (" js/klass ")"]
+   [notes-component]
+   ;; FIXME: 日付が今日で、まだサブミットしていない時に出す。
+   (when (today-is-klass-day? js/klass)
+     [new-note-compoment])])
 
 (def pages
   {:home #'home-page
@@ -57,8 +91,8 @@
 
 (def router
   (reitit/router
-    [["/" :home]
-     ["/about" :about]]))
+   [["/" :home]
+    ["/about" :about]]))
 
 (defn match-route [uri]
   (->> (or (not-empty (string/replace uri #"^.*#" "")) "/")
@@ -71,22 +105,18 @@
 (defn hook-browser-navigation! []
   (doto (History.)
     (events/listen
-      HistoryEventType/NAVIGATE
-      (fn [^js/Event.token event]
-        (swap! session assoc :page (match-route (.-token event)))))
+     HistoryEventType/NAVIGATE
+     (fn [^js/Event.token event]
+       (swap! session assoc :page (match-route (.-token event)))))
     (.setEnabled true)))
 
 ;; -------------------------
 ;; Initialize app
-(defn fetch-docs! []
-  (GET "/docs" {:handler #(swap! session assoc :docs %)}))
-
 (defn ^:dev/after-load mount-components []
   (rdom/render [#'navbar] (.getElementById js/document "navbar"))
   (rdom/render [#'page] (.getElementById js/document "app")))
 
 (defn init! []
   (ajax/load-interceptors!)
-  (fetch-docs!)
   (hook-browser-navigation!)
   (mount-components))